requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
docopt==0.6.2
gitignore==0.0.8
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.4
packaging==24.1
Pillow==9.5.0
promise==2.3
psycopg2==2.9.9
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.4.0
